% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grind.R
\name{font.main}
\alias{font.main}
\title{Font for plot titles (1 = plain, 2 = bold); see \code{\link[graphics]{par}}}
\usage{
font.main
}
\description{
Font for plot titles (1 = plain, 2 = bold); see \code{\link[graphics]{par}}
}
