% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grind.R
\name{font.sub}
\alias{font.sub}
\title{Font for plot subtitles (1 = plain, 2 = bold); see \code{\link[graphics]{par}}}
\usage{
font.sub
}
\description{
Font for plot subtitles (1 = plain, 2 = bold); see \code{\link[graphics]{par}}
}
