% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grinder}
\alias{grinder}
\title{Coffee grinder cup counts}
\format{A data frame with 29 rows and 2 variables:
\describe{
  \item{date}{the date of observation}
  \item{cups}{observed cup count}
}}
\usage{
grinder
}
\description{
A dataset containing the total cup count on a given day for the coffee grinder
at the Department of Computer Science, Univeristy of Tromsø. We have
gathered the observations by visual inspection of the digital display on
the grinder itself.
}
\keyword{datasets}

