% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grind.R
\docType{package}
\name{grindR-package}
\alias{grindR}
\alias{grindR-package}
\title{grindR: Phase plane analysis and parameter estimation for ODE models}
\description{
A wrapper around deSolve, rootSolve, and FME that provides five easy-to-use functions for analyzing ODE models: run() for numerical integration, plane() for phase plane analysis, newton() for finding steady states, continue() for bifurcation diagrams, and fit() for parameter estimation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tbb.bio.uu.nl/rdb/grind.html}
}

}
\author{
\strong{Maintainer}: Bram van Dijk \email{b.vandijk@uu.nl}

Authors:
\itemize{
  \item Bram van Dijk \email{b.vandijk@uu.nl}
  \item Rob J. de Boer \email{r.j.deboer@uu.nl}
}

}
\keyword{internal}
