% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_grind.R
\name{fit}
\alias{fit}
\title{fit}
\usage{
fit(datas = data, state = s, parms = p, odes = model,
  free = NULL, who = NULL, differ = NULL, fixed = NULL, tmin = 0,
  tmax = NULL, ymin = NULL, ymax = NULL, log = "", xlab = "Time",
  ylab = "Density", bootstrap = 0, show = NULL, fun = NULL,
  costfun = cost, initial = FALSE, add = FALSE, timeplot = TRUE,
  legend = TRUE, main = NULL, sub = NULL, pchMap = NULL, ...)
}
\description{
fit
}
