% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_grind.R
\name{newton}
\alias{newton}
\title{newton}
\usage{
newton(state = s, parms = p, odes = model, time = 0, x = 1,
  y = 2, positive = FALSE, jacobian = FALSE, vector = FALSE,
  plot = FALSE, ...)
}
\description{
newton
}
