% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_grind.R
\name{plane}
\alias{plane}
\title{plane}
\usage{
plane(xmin = 0, xmax = 1.1, ymin = 0, ymax = 1.1, log = "",
  npixels = 500, state = s, parms = p, odes = model, x = 1,
  y = 2, time = 0, grid = 5, eps = 0, show = NULL,
  portrait = FALSE, vector = FALSE, add = FALSE, legend = TRUE,
  zero = TRUE, lwd = 2, col = "black", pch = 20, ...)
}
\description{
plane
}
