% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_grind.R
\name{run}
\alias{run}
\title{run}
\usage{
run(tmax = 100, tstep = 1, state = s, parms = p, odes = model,
  ymin = 0, ymax = NULL, log = "", x = 1, y = 2, xlab = "Time",
  ylab = "Density", tmin = 0, draw = lines, times = NULL,
  show = NULL, arrest = NULL, after = NULL, tweak = NULL,
  timeplot = TRUE, traject = FALSE, table = FALSE, add = FALSE,
  legend = TRUE, solution = FALSE, delay = FALSE, lwd = 2,
  col = "black", pch = 20, ...)
}
\description{
run
}
