% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_grind.R
\name{timePlot}
\alias{timePlot}
\title{timePlot}
\usage{
timePlot(data, tmin = 0, tmax = NULL, ymin = 0, ymax = NULL,
  log = "", xlab = "Time", ylab = "Density", show = NULL,
  legend = TRUE, draw = lines, lwd = 2, add = FALSE, main = NULL,
  sub = NULL, recolor = FALSE, colMap = NULL, pchMap = NULL, ...)
}
\description{
timePlot
}
